refactor(referentiels): réorganise switch-to-te/shared en origine, cible et correspondance
Clarifie la dualité action-origine / action-cible et pré-calcule correspondanceIndexes dans le contexte pour éviter de reconstruire les index à chaque merge de liens fiches.
