+++ apps/backend/src/referentiels/switch-to-te/shared/action-origine.spec.ts
@@ -5,12 +5,16 @@ import {
   ReferentielIdEnum,
   type ReferentielId,
 } from '@tet/domain/referentiels';
+import { hierarchiesByReferentielIdForTests } from './referentiel-hierarchies.test-fixture';
 import {
+  collectMesureOrigineIds,
   dedupeOrigines,
   filterOriginesConcernees,
-  isCibleConcernee,
+  resolveMesureOrigineId,
   sortByReferentielOrder,
-} from './origine.rules';
+} from './action-origine';
+
+const hierarchies = hierarchiesByReferentielIdForTests;
 
 const createActionScore = (overrides: Partial<ActionScore> = {}): ActionScore =>
   ({
@@ -161,24 +165,88 @@ describe('sortByReferentielOrder', () => {
   });
 });
 
-describe('isCibleConcernee', () => {
-  it('retourne true si le score est absent', () => {
-    expect(isCibleConcernee(new Map(), 'te_1')).toBe(true);
+describe('resolveMesureOrigineId', () => {
+  it('remonte une tâche CAE vers la mesure source', () => {
+    expect(
+      resolveMesureOrigineId(
+        {
+          referentielId: ReferentielIdEnum.CAE,
+          actionId: 'cae_6.1.3.4.3',
+        },
+        hierarchies
+      )
+    ).toBe('cae_6.1.3');
   });
 
-  it('retourne true si concerne est true', () => {
-    const scoreMap = new Map<string, ActionScore>([
-      ['te_1', createActionScore({ concerne: true })],
-    ]);
-
-    expect(isCibleConcernee(scoreMap, 'te_1')).toBe(true);
+  it('remonte une tâche ECI vers la mesure source', () => {
+    expect(
+      resolveMesureOrigineId(
+        {
+          referentielId: ReferentielIdEnum.ECI,
+          actionId: 'eci_3.3.1.3',
+        },
+        hierarchies
+      )
+    ).toBe('eci_3.3');
   });
 
-  it('retourne false si concerne est false (personnalisation / désactivation)', () => {
-    const scoreMap = new Map<string, ActionScore>([
-      ['te_1', createActionScore({ concerne: false, pointPotentiel: 0 })],
-    ]);
+  it('remonte une sous-action ECI vers la mesure source', () => {
+    expect(
+      resolveMesureOrigineId(
+        {
+          referentielId: ReferentielIdEnum.ECI,
+          actionId: 'eci_3.3.1',
+        },
+        hierarchies
+      )
+    ).toBe('eci_3.3');
+  });
 
-    expect(isCibleConcernee(scoreMap, 'te_1')).toBe(false);
+  it('laisse inchangé une origine déjà au niveau mesure', () => {
+    expect(
+      resolveMesureOrigineId(
+        {
+          referentielId: ReferentielIdEnum.CAE,
+          actionId: 'cae_6.1.3',
+        },
+        hierarchies
+      )
+    ).toBe('cae_6.1.3');
+  });
+
+  it('retourne actionId inchangé si hiérarchie absente', () => {
+    expect(
+      resolveMesureOrigineId(
+        {
+          referentielId: ReferentielIdEnum.CAE,
+          actionId: 'cae_6.1.3.4.3',
+        },
+        new Map()
+      )
+    ).toBe('cae_6.1.3.4.3');
+  });
+});
+
+describe('collectMesureOrigineIds', () => {
+  it('agrège les mesures sources dédupliquées', () => {
+    const ids = collectMesureOrigineIds(
+      [
+        {
+          referentielId: ReferentielIdEnum.CAE,
+          actionId: 'cae_6.1.3.4.3',
+        },
+        {
+          referentielId: ReferentielIdEnum.CAE,
+          actionId: 'cae_6.1.3.4.1',
+        },
+        {
+          referentielId: ReferentielIdEnum.ECI,
+          actionId: 'eci_3.3.1.3',
+        },
+      ],
+      hierarchies
+    );
+
+    expect([...ids].sort()).toEqual(['cae_6.1.3', 'eci_3.3']);
   });
 });
